html/webappapis/dynamic-markup-insertion/document-write/module-dynamic-import.html

<!doctype html>
<meta charset=utf-8>
<title>document.write in an imported module</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(test => {
  const iframe = document.createElement("iframe");
  iframe.src = "module-dynamic-import-iframe.html";
  document.body.appendChild(iframe);
  iframe.onerror = test.unreached_func("Error loading iframe");

  let onLoadWasCalled = false;
  iframe.onload = test.step_func(() => {
    assert_equals(iframe.contentDocument.body.textContent, "FAIL\n");
    iframe.onload = null;
    onLoadWasCalled = true;
  });
  document.addEventListener("testEnd", test.step_func_done(() => {
    assert_true(onLoadWasCalled);
    assert_equals(iframe.onload, null);
    assert_equals(iframe.contentDocument.body.textContent, "PASS\n");
  }));
});
</script>
